feat: ARC, Memory Leaks, and Capture Lists [weak self]
